update user repository方法

+++ DTOs/Response/UserDto.cs
@@ -2,7 +2,7 @@
 {
     public sealed class UserDto
     {
-        public int Id { get; init; } = default!;
+        public long Id { get; init; } = default!;
         public string? Email { get; init; }
         public string? DisplayName { get; init; }
     }
